requirements.txt

langchain
langchain-core
langchain-groq
langchain-openai
langchain-google-genai
langchain-community
langgraph
langchain-tavily  
tavily-python


fastapi
pydantic
uvicorn[standard]


streamlit


python-dotenv
requests